requirements.txt
streamlit
chromadb
google-generativeai
python-dotenv
reportlab
st_audiorec